First Demo Mode implementation

+++ frontend/src/utils/dataApiConnector.ts
@@ -1,4 +1,5 @@
-import { Event, Participate, Profile, Project, STORAGE_PROFILE } from '../types/types';
+import { DEMO_RESULTS } from '../types/demoData';
+import { Event, Participate, Profile, Project, RoleTypes, STORAGE_PROFILE } from '../types/types';
 import axios from 'axios';
 
 export enum ResultType {
@@ -34,6 +35,11 @@ export const loadStoredProfile = (): Profile | null => {
   return userProfile;
 };
 
+export const isDemo = (profile: Profile): boolean => {
+  const nonDemoRoles = [RoleTypes.ADMIN, RoleTypes.MANAGER, RoleTypes.USER];
+  return !profile || !profile.role_id || !nonDemoRoles.includes(profile.role_id);
+};
+
 /* *************************************************** */
 /* User API                                            */
 /* *************************************************** */
@@ -49,7 +55,7 @@ export const getProfile = async (
     });
     return resultSuccess(response.data);
   } catch (error) {
-    return resultError('Profile konnte nicht geladen werden');
+    return resultError('getProfile: Profil konnte nicht geladen werden');
   }
 };
 
@@ -69,7 +75,7 @@ export const putProfile = async (
       return resultError('Profil konnte nicht aktualisiert werden');
     }
   } catch (error) {
-    return resultError('Profil konnte nicht aktualisiert werden');
+    return resultError('putProfile: Profil konnte nicht aktualisiert werden');
   }
 };
 
@@ -98,6 +104,7 @@ export const deleteProfile = async (
   token: string | null,
 ): Promise<{ resultType: ResultType; resultMsg: string | null; data: Profile | null }> => {
   if (!token || !profile) return resultError('token or profile is missing');
+  if (isDemo(profile)) return resultError('deleteProfile: Not allowed in demo mode');
 
   try {
     const response = await axios.delete(`/api/user/${profile.id}`, {
@@ -173,9 +180,17 @@ export const postActivate = async (
 /* *************************************************** */
 export const getProjects = async (
   eventId: number | null,
+  profile: Profile | null = null,
   token: string | null,
 ): Promise<{ resultType: ResultType; resultMsg: string | null; data: Project[] | null }> => {
-  if (!token || !eventId) return resultError('token or userId is missing');
+  if (!token || !eventId) return resultError('token or eventId is missing');
+  if (!profile) return resultError('profile is missing');
+  if (isDemo(profile)) {
+    console.log('getProjects: Using DEMO data');
+    console.log('getProjects: eventId:', eventId);
+    console.log('getProjects: DEMO_RESULTS.projects:', DEMO_RESULTS.projects[eventId]);
+    return resultSuccess(DEMO_RESULTS.projects[eventId]);
+  }
 
   try {
     const response = await axios.get<Project[]>(`/api/project/list/${eventId}`, {
@@ -184,7 +199,7 @@ export const getProjects = async (
     return resultSuccess(response.data);
   } catch (error: any) {
     if (error?.response?.status === 404) return resultSuccess([]);
-    return resultError('Projects konnten nicht geladen werden');
+    return resultError('getProjects: Projects konnten nicht geladen werden');
   }
 };
 
@@ -233,8 +248,10 @@ export const postProject = async (
 /* *************************************************** */
 export const getEvents = async (
   token: string | null,
+  profile: Profile | null = null,
 ): Promise<{ resultType: ResultType; resultMsg: string | null; data: Event[] | null }> => {
-  if (!token) return resultError('token  is missing');
+  if (!token || !profile) return resultError('token or profile is missing');
+  if (isDemo(profile)) return resultSuccess(DEMO_RESULTS.events);
 
   try {
     const response = await axios.get<Event[]>(`/api/event/list`, {
@@ -247,22 +264,22 @@ export const getEvents = async (
   }
 };
 
-export const getEvent = async (
-  id: number | null,
-  token: string | null,
-): Promise<{ resultType: ResultType; resultMsg: string | null; data: Event | null }> => {
-  if (!token || !id) return resultError('token or id is missing');
+// export const getEvent = async (
+//   id: number | null,
+//   token: string | null,
+// ): Promise<{ resultType: ResultType; resultMsg: string | null; data: Event | null }> => {
+//   if (!token || !id) return resultError('token or id is missing');
 
-  try {
-    const response = await axios.get<Event>(`/api/event/${id}`, {
-      headers: { Authorization: `Bearer ${token}` },
-    });
-    return resultSuccess(response.data);
-  } catch (error: any) {
-    if (error?.response?.status === 404) return resultSuccess(null);
-    return resultError('Event konnten nicht geladen werden');
-  }
-};
+//   try {
+//     const response = await axios.get<Event>(`/api/event/${id}`, {
+//       headers: { Authorization: `Bearer ${token}` },
+//     });
+//     return resultSuccess(response.data);
+//   } catch (error: any) {
+//     if (error?.response?.status === 404) return resultSuccess(null);
+//     return resultError('Event konnten nicht geladen werden');
+//   }
+// };
 
 /* *************************************************** */
 /* Participate                                         */
